third_party/blink/web_tests/platform/mac/external/wpt/mathml/presentation-markup/operators/display-operator-min-height-default-font-expected.txt

This is a testharness.js-based test.
[FAIL] Operators using largeop should be bigger
  assert_greater_than: Test ∫ expected a number greater than 19.51875 but got 16.265625
Harness: the test ran to completion.
